Change the package structure to support new features and goals
All toybox packages are now self-contained entities capable of installing themselves into and uninstalling themselves from the target system. Look to toybox.toy for an example of the new package structure.

The toybox script now allows multiple packages to be specified following the install or uninstall options (and so also supports wildcards), but is otherwise backward compatable with the previous version.

As requested packages may now be installed or uninstalled from any location on the system by specifying the relative path to the package.